opengl: interop: change signature
Provides the vlc_gl_interop object as parameter to the activate function
for the modules, instead of a generic vlc_object_t. This removes the
need for casting the vlc_object_t and clarify how interop modules are
loaded. The new activation function will also allow to supply new
parameters in the future.