qt: MainInterface: decode partially encoded MRL on drop
When using drag&drop to play a media with characters that would need
encoding, especially spaces, the system will provide the correct URL to
the DropArea but Qt will encode it partially. Since we will encode it
afterwards, we can just decode it completely and forward it as-is to the
playlist controller.

Fixes dropping URL with spaces from a browser to VLC.