image: separate the encoding codec from the output video format
The output is a bitstream in some codec, not a chroma that can be used as such.
The codec value should not be set as a chroma in the video_format_t.

The output video format is only used to force some dimensions on the output
that may not be the ones from the source.