direct3d11: remove padding in the pool picture
On 3.0 the pool video format doesn't have to match the decoder actual
padding. The quad handling to place the picture will take care of this
with the actual texture size.

We still need to fix the semiplanar dimensions although it's unlikely
the quad will use anything else than an RGB output.

Fixes green line seen in fullscreen.