input_clock: remove DVD CR gap check
The check was added back then for a special case with DVDs, but it
should be moved to the DVD demuxer instead if it is really needed.
However, it actually request the system time to have a min value which
is almost always the case now, so it's probably not needed anymore.