clock: don't include the delay for the master clock
If the source is the master, it needs to handle the delay itself by
delaying the playback by the delta returned by vlc_clock_SetDelay() (for
slaves, the return value will be always 0). This is already done by the
audio output, that is sending silence to reach that delta.

The delay was applied 2 times, one time in the audio output and one time
in the clock. Therefore, remove that extra delay.