sout: duplicate: forward a correct PCR to sink
duplicate modules can sink back to the main pipeline which was causing
duplicated and inconsistent PCR in the main pipeline.
An example duplicate module with a sink:
`#duplicate{dst=...,dst=...}:std{...}`

To fix the issue this patch add an endpoint to each duplicated sout
chains before they sink back to the main pipeline. We use it to regroup
PCR from each chains and select the lowest to forward it to the sink.