input_clock: export clock listener
Use a generic clock listener instead of the vlc_clock_t based-one so
that we can attach a different listener to the input clock, in
particular in the case of tests.

The change moves some code back to the es_out.c file, which we try to
simplify, but the creation of the clock and buffering handling as a
whole might be moved somewhere else someday.