sout: introduce a close callback
Adding the close callback in the object v-table instead of relying on
the module to do so.
This remove module dependency from the stream output object. From now
on, it will be possible to allocate and set-up stream outputs
programmatically without having to declare a module.

This is useful in scenario where a very contextual stream output filter
implementation is needed and we don't want to expose a module that the
user can manipulate. A good example of that is the current
`chromecast-proxy` submodule which inherit from it's parent internals
and drop data until the first I-Frame.

This approach of delegating object deletion to the v-table is already
taken by most modern module types.