sout: allow module-less sout creation
With the new `close` callback exposed in the v-table it becomes possible
to create empty sout objects. Those allow to programmatically create
custom stream outputs without having to declare them as shared modules.

This will be used to replace the current `chromecast-proxy`
implementation or to forward PCR properly in duplicate (we need specific
sout filters here to catch PCR at the end of each duplicated stream).