sout: sdi: move sout callbacks out of the main class
Let's switch to the ops initializer used in most C++ modules. This is
preparation work to switch to the sout close callback in ops. The close
callback in this context cannot be owned by the SDIOutput class as it
needs to have a complete definition of the object implementing this
interface.