subpicture: remove subpicture_region_Copy()
It's not used anymore.

As the doc suggests it was provided for convenience and was originally not
handling text-based regions at all. Is it supposed to take care of the
subpicture_region_t::p_private data copying it doesn't know about ?

It is rare to need this feature on regions. In bluray it was likely done
because the same region couldn't be in 2 different containers at the same time.