encoders: fill missing input format video chroma
If an encoder requires a specific input chroma, it needs to be specified
in both the codec and the chroma field of the input format. The encoder
creators currently fix the input format themselves when an encoder does
not provide any chroma.

This input format fix was legitimately removed from VLC core's image API
which was causing regressions with the chroma converter attachment.