sout: mosaic_bridge: always create the filter chain
Always creating the filter chain lighten the nesting and avoids a lot of
NULL checking. Manipulating empty filter chains if the user does not
specify any filter is fine.

This is preparation work to actually use the filter chain to do the
chroma resampling and rescaling instead of doing it manually in this
module.