sout: mosaic_bridge: move decoder owned var to owner
Those variable have the same lifetime as the decoder owner or should
only be used when its created.
For var callbacks, It requires taking the global lock but, for most of
those variable callbacks where setting values read by other modules
without those being atomic so it is likely that taking the lock was
necessary to avoid potential race conditions anyway.