sout: mosaic_bridge: convert pictures with the filter chain
Instead of handling the scaling and chroma conversion separately via the
image API, prepend the already existing filter chain with a converter.
This gets closer to what we have in transcode and remove the needs for
the picture API here. It also avoids computing the scaling values at
every frame queued.

Variable callbacks needed to be adapted as the filter chain has to be
recreated at the decoder queue level for that use case. A local lock is
introduced to avoid taking advantage of the global lock just to set
variables that are local to the module but still need locking for the
owner access.