qml: normalize drop shadow drawing
image provider requested size is dpr dependant, so drawing needs to be scaled
accordingly. All metrics provided to draw the shadow are provided in the same
referencial which is dpr independant (usually Qml referential). This referential
is expressed through the "viewport" variables which indicates the bounds of the
picture to draw.

for the effective drawing, The viewport is scaled  fit in the requested image
size, then we draw the effect by applying the computed scale factor to all
metrics.

One other benefit of this, is that we can limit the CPU & memory required for
the drawing by limitting the image sourceSize