clock: continue to convert when paused
The paused state is already sent to all outputs, outputs might want to
convert timestamps while paused.

This is the case for the SPU output that need to select subpictures to
be displayed while paused. The SPU output still won't select the next
subpictures, since the vout will be paused and will always try to play
the same picture with the same converted TS.

Fixes #28429