qt: Makefile.am: use recursive Makefile
The project started to move away from recursive Makefile a while ago for
performance reasons. Unfortunately, even though tools like libtool can
handle an arbitrary number of files by using @filelists, automake is
subject[^1][^2] to max argument limitation in its `distdir-am` target,
which will try to expand $(DISTFILES) in the bash command line.

[^1]: https://lists.gnu.org/archive/html/bug-automake/2009-07/msg00008.html
[^2]: https://lists.gnu.org/archive/html/automake/2014-05/msg00015.html

This is workarounded in this commit by moving all the files from Qt in a
separate Makefile that will be recursively called by automake, reducing
the size of $(DISTFILES). Hopefully, Make/Automake would have fixed this
or we would have moved completely to Meson when we reach the limit
again.