qt: Makefile.am: expose each test separately
The test was running every QML test files from the folder, which means
that we could have tests running with `make check` but not on the
repository. Test results would also be mixed up together in the same
files.

Since a QML test can run multiple tests itself, it's closer to the other
native tests from the repository.