qt: Makefile.am: symlink vlc-qt-check to modules/
vlc-qt-check is found using VLC_PKG_LIBEXEC_DIR which falls back to
VLC_LIB_PATH. In vlc-static, VLC_LIB_PATH is defined by the following
snippet:

    #ifdef TOP_BUILDDIR
        setenv ("VLC_PLUGIN_PATH", TOP_BUILDDIR"/modules", 1);
        setenv ("VLC_DATA_PATH", TOP_SRCDIR"/share", 1);
        setenv ("VLC_LIB_PATH", TOP_BUILDDIR"/modules", 1);
    #endif

With the previous changes in ddfc6d0de8ae0f450895ec5cee6626e4efacb422,
the vlc-qt-check executable is not built into TOP_BUILDDIR/modules/
anymore, but in TOP_BUILDDIR/modules/gui/qt/ instead, and won't be found
on platforms requiring it.

When installing, it is still located at the right place, but this commit
adds a symbolic link to modules/ to ensure we can also find it with
vlc-static.