sout: hls: move segmentation to access output
The segmentation was previously done in SetPCR to output segments at
precise stream times instead of reliying on the quantity of data output
by the muxers.

This had two main issues:
- SetPCR does not report errors and the segment creation phase can fail
  at multiple stages.
- The calculated stream time did not actually reflect the quantity of
  data that left muxers. Muxers can be delayed by mux caching and are
  allowed to have an internal queue. The previous algorithm was
  outputting smaller segments due to the muxer delay.

Those issues were fine with TS as a single muxer but with the future
introduction of VTT/FMP4 segmentation that both make choice base on
segment times, we needed to change the segmenting strategy.

Segmentation is now done at muxer output, bigger segments are output
when all the muxers sent the right amount of data.