sout: hls: add webvtt support
WebVTT and plain subtitle tracks can now be exposed in HLS. an extra
media rendition is created for each subtitle ES and the HLS server
creates max-sized vtt segments following the pace of the other ES.

WebVTT segmentation in scenario where no subtitle is output for a while
is handled using the PCR of the stream. Having reliable clock info helps
a lot to create empty subtitle segments at the correct time.