libaribcaption: remove unneeded strlen() check
var_InheritString() guarantees that, if the string returned is not NULL,
the string does not contain a NULL terminator as first character.

Checking for strlen() > 0 does the same as checking if the string
contains a NULL terminator as first character, because strlen() doesn't
count the NULL terminator. Thus, remove the unneeded check.