update: update to 64-bit VLC on 64-bit Windows
This is justified just by the sheer amount of 64-bit arithmetic we do.

The only point in sticking to a 32-bit VLC is binary compatibility,
either for external VLC plugins, or for LibVLC applications. But VLC 4.0
will break binary compatibility on both counts, so that is moot.
This constitutes a once in a decade opportunity to smoothly switch over,
considering how rarely LibVLC compatibility is broken.