core: move Internal{Dialog,Keystore}Clean
In some cases, when a module uses a dialog, the dialog provider is
destroyed before the module is unloaded. This can lead to a crash if
the module continues to use it after it has been released.

This is fixed by moving InternalDialogClean after the vlc_playlist_Delete
so that it remains available during the player cleanup and the input
thread cleanup.