flac: fix endless bytes skip
A hang was caused when the stream head is invalid and a header should be skipped, but not enough data is available. In this case, block_SkipBytes returns VLC_EGENERIC and does not seek anything. So, the same invalid block is processed again and again.

(cherry picked from commit 306505b7c460a7078bf2ce481dd8c5d866c6f8c7)