demux/mkv: define the file context with no globals like the real one
We should not access the internal list of the original context directly. The type will change.

In a Matroska or WebM file there can only be an Ebml Header or a Matroska at the top level.
The CRC32 cannot be found. And the Void element should not be found either.