access/vnc: fix endianness handling
Outside of keeping bits per pixel consistent, we never made any attempt
to match the server's native pixel format in the first place, always
requesting our own hardcoded format for any color depth, so we might as
well also always ask the server for LE byte order (regardless of client
endianness) and code around that.  This change has been tested with LE
and BE clients.  Previous to this commit colors were wrong in the case
of (e.g.) LE <-> LE clients and servers.