access/vnc: set trueColour flag
In the unlikely event that the server's native pixel format is not true
color, this flag will be received as zero, and so we cannot assume that
it has already been set.  Since we only support true color pixel
formats, we therefore need to ensure that the flag is set ourselves.