input_clock: expose drift in update callback
input_clock_Update() is returning a value representing whether too much
buffer has been consumed or not, but it doesn't use any external value
to compute it.

After the update, the input_clock is running UpdateListener() to update
any code that would use the value reported by the input buffering, and
specifically the vlc_clock_t when running inside the es_out. This was
triggering a vlc_clock_Update().

vlc_clock_Update() returns the drift of the updated clock compared to
the current playback time. It wasn't used for the input clock because
this drift is always zero when the updated clock is driving the clock
bus ("master clock") and the vlc_clock_t was not added as a listener
when the input clock was not driving the clock bus.

The drift value will be reported from the clock update to the input
clock. It currently has no usage except tracing, but the usage will
replace the i_late computation in the es_out.c in the end.