es_out: always create an input vlc_clock_t
When the input is not selected to drive the clock bus, no vlc_clock_t
was assigned to track the drift of the input when compared to the audio
clock or other clocks, meaning that there was no way to inspect the
behaviour of the input buffering when not master.

This commit refactors the input vlc_clock_t creation to always have a
vlc_clock_t and always update it.