include: vlc_list.hpp: add insertion/removal functions
Add some methods to replace their C counterpart when using a vlc_list,
so that elements can be added and removed. Those methods are only
supported on a vlc::list<T> wrapper, but they are not supported neither
on the vlc::const_list<T> wrapper (which is expected) nor on the
vlc::list::reverse_list wrapper from ::as_reverse() function, because it
would complexify the design.

The current workaround that was chosen is that vlc::list<T> is able to
use iterators from the reversed list, and can be swapped with the
reverse list object where needed.

The limitation will be lifted with C++20 and concepts.

Co-authored-by: Alexandre Janniaux <ajanni@videolabs.io>