src: test: add test for vlc::list<> wrappers
The tests will check the correct behaviour of the iterators as well as
the type validity for the templates once they are instantiated.

They also check whether removal of the current element during iteration
is working correctly both on the normal list wrapper and the reversed
version. This is important given how tricky std::reverse_iterator can
be.

Co-authored-by: Pierre Lamot <pierre@videolabs.io>