src: Makefile.am: split C and C++ headers
Separate .h and .hpp headers in the pluginsinclude_HEADER variable to be
able to associate separate target variables for them, in particular for
the checkheader feature.

Indeed, POSIX make implementations are perfectly able to create macro
substitution and merge multiple macros together, but they are not able
to filter or substract in a general way (only one-item substraction
works) from an existing macro, so the way this should be done is by
splitting the source items and merge them where needed.