src: Makefile.am: fix compatiblity with .hpp
The checkheaders mechanism wasn't compatible with .hpp headers, and even
more, because of substitution failure, the .hpp headers were deleted
when make distclean was run.

This commit uses the newly-separated pluginsinclude_HEADERS.h and
pluginsinclude_HEADERS.hpp variables to supply different variants for
the CHECKCOMPILE and CHECK_SOURCE_EXT variable depending on the file
type.