contrib: amf: update to 1.4.30 using git
It's faster to get the files from git and compress than downloading
the 400 MB+ package.

We don't use download_git which downloads the whole git history of the
given branch. We use git clone and a sparse checkout to only get the
folder we want for the giant repository.

(cherry picked from commit 7cfdd439b9ba299606456a62de15b144201176c9)