extras: ci: enable coverage reports in MR
To enable coverage report in MRs (and not just in artifacts), we need to
indicate the regex for gitlab to parse the total percentage. See general
documentation[^1] highlighting that uploading doesn't enable, as well as
how to enable history [^2] and in merge requests[^3] for more details.

The || true is needed on the script line extracting the coverage
percentage, because the job is failing otherwise, but I'm still not sure
what is causing the failure.

[^1]: https://docs.gitlab.com/ee/ci/testing/test_coverage_visualization.html
[^2]: https://docs.gitlab.com/ee/ci/testing/code_coverage.html#view-history-of-project-code-coverage
[^3]: https://docs.gitlab.com/ee/ci/testing/code_coverage.html#view-code-coverage-results-in-the-mr

Co-authored-by: Alaric Senat <dev.asenat@posteo.net>