extras: ci: fix get-contrib-sha for debian/android
Running

    ./extras/ci/get-contrib-sha.sh debian-meson

is currently failing with:

    fatal: empty string is not a valid pathspec. please use . instead if
    you meant to match all paths

because "${VLC_BUILD_TARGET}" would count for an empty parameter, as
opposed to being discarded without quotes. In some shells, like zsh, it
would also count as a parameter regardless of the presence of the
quotes.

Instead of using an intermediate variable to store an optional path,
this commit directly adds to the bash array containing the path checked
to find the contrib ancestor, allowing to both submit multiple files if
necessary, or none if unnecessary.

Regression from 4ac6168d4740b6416fa6abbff30b50b14d370ef9.