test: meson: disable test with disabled module dependencies
Some tests will require some modules to be present to run correctly, but
it cannot be probed at compile time. However, as soon as every modules
will be declared unconditionnally, we'll be able to depend on a disabled
module and disable the matching tests.

Ideally, we'd skip it so that the test is still reported, but there's no
code to do that yet.