module: meson: always fetch frameworks
Since the dependencies are declared conditionnally, it was not possible
to unconditionnally create plugins that would depend on them. By always
declaring them, they are now defined as a disabler() when the condition
are not met, without even the need to probe them.

In addition, the setup factors together the conditions that are needed
to probe or not the dependencies, and put everything in a single
location that can be shared between the different Apple platforms. It
also makes it clearer what kind of dependency is linked, though meson
was already abstracting all of the framework handling, and what is the
real name that should be found in the documentation.