es_out: reset main clock with input clock
Currently, the main clock is reset once the new buffering is done. This
buffering step will happen at the following locations:
 - at the beginning of a playback
 - when a seek is triggered by the user
 - when the demux needs to trigger RESET_PCR

But the input clock was already reset at this location and will start to
provide new points. Resetting the clock there as well as when the
buffering is done currently changes nothing because we either have
already flushed everything (seek or RESET_PCR) or we are not running
through a clock (frame-by-frame mode).

However, future patches will remove the clock reset at the end of the
buffering and expect the clock to be already reset, to introduce a start
function for the clock.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>