contrib: opus: Simplify a condition for Windows on aarch64
While the CI configuration where this is encountered is LLVM specific
(because there is no upstream GCC that targets Windows on aarch64),
the issue that Opus lacks a RTCD implementation for Windows on
AArch64; this is not toolchain specific. Thus remove the unnecessary
HAVE_CLANG condition.

Also simplify the naming of the OS/arch combination (even if the
old naming presumably was chosen to match the CI configuration).