qml: use qsTr() instead of I18n.qtr() for translation
I18n module was intending to provide translation through vlc_gettext().
I'm not sure why QTranslator was not the way chosen before, because it
directly offers this and using qsTr() has the advantage of better integration.

Translation is handled in a lower abstraction level through QJSEngine
(see QJSEngine::TranslationExtension) as opposed to using custom a QML type
through QQmlEngine.

Thanks to that, the `org.videolan.vlc` module import can be eliminated
in some files.