vout_subpictures: rework the persistent cache of subtitle positions
This was forcing the subtitle region to become absolute. This will
later break subtitles in black bar which can be outside of the video
area. Absolute subtitle areas can only describe a position in the video
area.

This was also broken if the subpicture updater decides to set the region
position at each subpicture update. It would update the forced location
we set earlier.