qml: use rectangle instead of image in PlayButton.qml
Originally, the interface had view transition
animations and they interfered with the appearance
of the play button.

There was a period in which opacity mask was used
to eliminate this issue, but it was removed because
opacity effect is expensive and should not be used
in a delegate. Alternatively, layering could be
used although was not used before.

Up until now, a svg image was used for the play
button background. An image has the advantage of
getting rendered in batch, provided that it is
placed in the atlas texture. However, image has
the disadvantage that arises from the cost
texture sampling incurs.

Since we do not use animations anymore, we can
simply use rectangle again.