contrib: qt: build qt modules with CMake
The scripts call just translate old options into their CMake counterpart.
The BUILD and BUILDPREFIX qt toolchain point to different build configurations.

The qt.toolchain.cmake corresponds to the toolchain file used by qt-cmake.
But we can't use it as it's supposed to be a tool for the cross-compiled
target. This toolchain file is the official one to use when cross-compiling Qt6
for Raspberry Pi in Qt Creator: https://wiki.qt.io/Cross-Compile_Qt_6_for_Raspberry_Pi