contrib: qtdeclarative: Add a dependency from qtdeclarative-tools to qtshadertools-tools
When cross building, and building the native qtdeclarative-tools
(which can start before building the main qtdeclarative),
these depend on tools installed by qtshadertools-tools (the
native build host version of qtshadertools).

This was already a dependency for the main, cross compiled
qtdeclarative target, but also is required for the qtdeclarative
native tools.