es_out_timeshift: use downcast class as parameter
The upper class was used everywhere, leading to the pointer being
downcasted at the beginning of every function. By using the downcast
class everywhere, we don't introduce explicit downcast in most places.

This commit also reduces the usage of the es_out_t type as parameter,
which will be moved to another downclass in follow-up MR.