contrib: meson-machinefile provide the binary from makefile
Provide the necessary binary overrides throught the machine files
exposing the contribs, and ensure that the meson-machinefile built
depends on which binaries are built.

The new build tool generation uses a new PKGS.tools variable which
should match with the names used to find the different programs.

Not all programs can be found this way. For instance, it seems that the
Qt module from meson will completely ignore this parameter and load for
the Qt installation anyway.