package/win32: remove the contrib target tools from the PATH
The cross-compiled binaries are likely not going to be usable on the host and
should not be added to the PATH. They should never be used to be consistent
between native and cross-compiled builds.

It was added in the original build script: 5648ecad1a15a75e2d71d0fa22dd9b1797b89431.