configure: keep the fxc path
We need to add it to the PATH when calling qsb which assumes it's in the PATH.
qsb doesn't have an option to tell it where to look (in contribs). So if it's
not in the PATH it will never find it.

Then we can use the one from contribs.