adaptive: FakeESOut: switch to lambda initialization
Switching to a zero-initialized structure provides the same behaviour as
designated initializers in C code, ie. optional callbacks don't need to
be specified during initialization. This simplify a lot adding and
removing callbacks to the virtual tables since we don't need to modify
every files using the table at the same time as changing the table
itself.