qml: establish binding in `delegateItem` property of FadingEdgeForListView
When `listView.itemAtIndex(0)` is used, the delegate item is
not resolved when new items added which makes it remain
null.

When `listView.count` is referred as done here, the QML
engine establishes a relational binding and this causes
the property to be updated each time the count is updated.

Regression since 300b3b9c.