qt: handle scene graph error signal for graceful failure
Scene graph initialization occurs asynchronously, while `makeMainInterface()`
expects an immediate return with an error code.

In other compositor integrations, it is trivial to return immediately. However,
in CompositorDirectComposition we have to wait for scene graph to initialize so
that we know what graphics API it is using because DComp composition can only work
with D3D11 and D3D12.

In order to satisfy that, an event loop is used here to wait until the scene
graph is initialized. This has worked fine, however there was no handling
in case an error occurs during scene graph initialization.

Now, we handle the error so that the interface is not stuck in waiting for
the event loop to quit. This will make the interface hence the Qt module
close gracefully.