player: timer: fix paused state overridden
Player timer states are not really states but events. It is possible to
receive a discontinuity while paused, in that case, the paused state was
overridden causing UI timers to continue while paused. Store each events
in different variables to fix this issue. Keeping the discontinuity
event in the internal player state is useless since
1ecb283c87865d01b5f3c947c0e0912f63011d64.

Fixes #28622